[IMP] im_livechat: remove AbstractThreadWindow
And also rename WebsiteLivechatWindow to PublicLivechatWindow

Task-2893507
